Update to libmad-0.15.1b-10.
Properly check the size of the main data. The previous change
only checked that it could fit in the buffer, but didn't ensure there
was actually enough room free in the buffer. This was assigned both
CVE-2017-8372 and CVE-2017-8373, but they are really the same, just a
different way to detect it.
Rewrite buffer size check. Now checks it before reading it instead
of afterwards checking that we did read too much. This now also covers
parsing the frame and layer3, not just layer 1 and 2.  CVE-2017-8374
mentions a case in layer 3.